+++ R/glossary_options.R
@@ -58,22 +58,47 @@ glossary_reset <- function() {
 #' Set or get the default glossary path
 #'
 #' @param path the path to the glossary file
+#' @param create create a new glossary file if it doesn't exist
 #'
 #' @return path string if path is NULL
 #' @export
 #'
 #' @examples
 #' glossary_path() # get current path
-glossary_path <- function(path = NULL) {
+#'
+#' \dontrun{
+#' # set path to glossary.yml file (assumes it exists)
+#' glossary("glossary.yml")
+#'
+#' # create (if doesn't exist) and set path
+#' glossary("include/glossary.yml", create = TRUE)
+#' }
+glossary_path <- function(path, create = FALSE) {
   # get path
-  if (is.null(path)) {
+  if (missing(path)) {
     return(glossary_options("path"))
   }
 
   # create file if necessary
-  if (path != "psyteachr" & !file.exists(path)) {
+  if (create &&
+      !is.null(path) &&
+      path != "psyteachr" &&
+      !file.exists(path)) {
+
+    # create directory if doesn't exist
+    dir <- dirname(path)
+    if (!dir.exists(dir)) {
+      dir.create(dir, FALSE, TRUE)
+    }
+
     write("term: |\n  definition\n", path)
-    message(path, " does not exist; it has been created")
+    message(path, " did not exist; it has been created")
+  }
+
+  if (!is.null(path) &&
+      path != "psyteachr" &&
+      !file.exists(path)) {
+    stop("The file ", path, " does not exist")
   }
 
   # set path
@@ -90,9 +115,9 @@ glossary_path <- function(path = NULL) {
 #'
 #' @examples
 #' glossary_popup() # get current popup style
-glossary_popup <- function(popup = NULL) {
+glossary_popup <- function(popup) {
   # get popup
-  if (is.null(popup)) {
+  if (missing(popup)) {
     return(glossary_options("popup"))
   }