+++ NEWS.md
@@ -1,3 +1,10 @@
+# glossary 0.0.0.9002
+
+* Cleaned up popup display of definitions with lists or line breaks
+* Fixed bug when setting `glossary_path(NULL)`
+* Added `create` argument to `glossary_path()` to control whether a new file should be created if it doesn't exist
+* Fixed bug in link to psyteachr glossary
+
 # glossary 0.0.0.9001
 
 * Added a `NEWS.md` file to track changes to the package.